trainer/requirements.txt
xgboost
numpy
joblib
google-cloud-storage
scikit-learn